Support more than 10 cores
The original implementation used a dirty hack to avoid having to piece
all filenames together individually: in a template string we overwrote
a single character with the (maximum one digit) core number.

The easy way would be generating all the file names individually, but
instead we create a number of templates for the number of digits needed
and keep the trick. But we encapsulate it away and even persist the
template strings.

The encapsulation required to have a static function reading the
configuration and passing on the data to a simpler constructor.